% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_chains.R
\name{append_chains}
\alias{append_chains}
\title{Append MCMC chains (objects of class \link[coda:mcmc]{coda::mcmc})}
\usage{
append_chains(...)
}
\arguments{
\item{...}{A list of \code{mcmc} or \code{mcmc.list} class objects.}
}
\value{
If \code{mcmc.list}, an object of class \code{mcmc.list}, otherwise,
an object of class \code{mcmc}.
}
\description{
Combines two or more MCMC runs into a single run. If runs have
multiple chains, it will check that all have the same number of chains, and
it will join chains using the \link{rbind} function.
}
