% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_mcmc.R
\docType{data}
\name{fmcmc-deprecated}
\alias{fmcmc-deprecated}
\alias{last_elapsed}
\alias{LAST_MCMC}
\alias{last_nsteps}
\alias{last_nchains}
\alias{last_kernel}
\alias{last_conv_checker}
\alias{last_}
\title{Deprecated methods in fmcmc}
\format{
An object of class \code{fmcmc_last_mcmc} of length 0.
}
\usage{
last_elapsed()

LAST_MCMC

last_nsteps()

last_nchains()

last_kernel()

last_conv_checker()

last_(x)
}
\arguments{
\item{x}{Name of the object to retrieve.}
}
\value{
The function \code{last_elapsed} returns the elapsed time of the last call
to \link{MCMC}. In particular, the \code{MCMC} function records the running time of R
at the beginning and end of the function using \code{\link[=proc.time]{proc.time()}}. So this function
returns the difference between the two (\code{time_end - time_start}).
}
\description{
These functions will no longer be included starting version 0.6-0. Instead,
use the functions in \link{mcmc-info}.
}
\keyword{datasets}
