% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\docType{data}
\name{fmcmc-info}
\alias{fmcmc-info}
\alias{MCMC_INFO}
\alias{get_}
\alias{get_logpost}
\alias{get_elapsed}
\alias{get_nsteps}
\alias{get_nchains}
\alias{get_kernel}
\alias{get_conv_checker}
\alias{get_seed}
\alias{get_burnin}
\alias{get_thin}
\title{Information about the last \code{MCMC} call}
\format{
An object of class \code{fmcmc_last_mcmc} (inherits from \code{fmcmc_info}, \code{environment}) of length 7.
}
\usage{
MCMC_INFO

get_(x)

get_logpost()

get_elapsed()

get_nsteps()

get_nchains()

get_kernel()

get_conv_checker()

get_seed()

get_burnin()

get_thin()
}
\arguments{
\item{x}{Character scalar. Name of an argument to retrieve. If \code{x} was not
passed to the last call, the function returns with an error.}
}
\value{
\verb{get_*} returns the corresponding variable passed to the last call
of \link{MCMC}.
}
\description{
This environment holds a copy of the last call to \link{MCMC}, including the start
and end time (to compute total elapsed time) of the call. Since the resulting
object of \code{MCMC} is an object of class \link[coda:mcmc]{coda::mcmc}, this is a way to capture
more information in case the user needs it.
}
\details{
The function \code{get_logpost} returns the \code{logposterior} value at each
iteration. The values correspond to a named numeric vector. If \code{nchains > 1}
then it will return a list of length \code{nchains} with the corresponding logpost
values for each chain.
}
\examples{
set.seed(23133)
x <- rnorm(200)
y <- x*2 + rnorm(200)
f <- function(p) {
  sum(dnorm(y - x*p, log = TRUE))
}

ans <- MCMC(fun = f, initial = c(0), nsteps=2000)
plot(get_logpost(), type = "l") # Plotting the logpost from the last run
}
\keyword{datasets}
