% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_sdml.r
\name{sdml_adaptive_kernel}
\alias{sdml_adaptive_kernel}
\title{State-Dependent Memory-Less Adaptive Transition Kernel}
\usage{
sdml_adaptive_kernel(x, f, gr, rz = stats::rnorm, rz.args = list(mean = 0,
  sd = 1), ..., alpha = 0.9, eps = rep(1e-05, length(x)))
}
\arguments{
\item{x}{Current state}

\item{f}{Objective function}

\item{gr}{(optional) Gradient of the objective function.}

\item{rz}{Random number function.}

\item{rz.args}{List of parameters passed to `rz`.}

\item{...}{Further arguments passed to `f` and, if specified, `gr`}

\item{alpha}{Number in `[0, 1)`.}

\item{eps}{If no `gr` is specified, epsilon used to approximate the gradient.}
}
\description{
State-Dependent Memory-Less Adaptive Transition Kernel
}
