% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autostop.R
\name{convergence-checker}
\alias{convergence-checker}
\alias{gelman_convergence}
\alias{geweke_convergence}
\title{Automatic stop}
\usage{
gelman_convergence(threshold = 1.1, check_invariant = TRUE)

geweke_convergence(threshold, ...)
}
\arguments{
\item{threshold}{Numeric value. A Gelman statistic below the threshold
will return \code{TRUE}.}

\item{check_invariant}{Logical. When \code{TRUE} the function only computes
the gelman diagnostic using variables with greater than \code{1e-10} variance.}

\item{...}{Further arguments passed to the method.}
}
\value{
A function passed to \link{MCMC} to check automatic convergence.
}
\description{
Automatic stop
}
\details{
In the case of \code{geweke_convergence}, \code{threshold} sets the p-value
for the null \eqn{H0: Z = 0}, i.e. equal means between the first and last
chunks of the chain. See \link[coda:geweke.diag]{coda::geweke.diag}. This implies that the higher
the threshold, the lower the probability of stopping the chain.

In the case that the chain has more than one parameter, the algorithm will
return true if and only if the test fails to reject the null for all the
parameters.
}
