% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autostop.R
\name{gelman_convergence}
\alias{gelman_convergence}
\title{Automatic stop}
\usage{
gelman_convergence(threshold = 1.1)
}
\arguments{
\item{threshold}{Numeric value. A Gelman statistic below the threshold
will return \code{TRUE}.}
}
\value{
A function passed to \link{MCMC} to check automatic convergence.
}
\description{
Automatic stop
}
